Move setup_workflow() test function in test utils
